Refactored BaseTeamplteIntegrationTestCase into trait
